socket.io implementation
The client can establish a connection with the server. Socketio runs on port 9092. Messages have to have the correct event name in order to be received. (see example in the code). Messages that are being sent to the server should include the type (client), a room (namespace) and a message. For this implementation i chose a default room 'a'. Currently, the clîent joins a room as soon as it connects. When a message is sent, the server broadcasts the message back to all membes of the room, including the sender. This can be adapted, of course.