member list is updated upon new join & some cleaning
The member list updates automatically. However, if you refresh the page, it will set the list to the state it was when the player, whose window is open, joined.
Also, a player can currently enter the gameRoom twice, unless it is the room leader. I haven't fixed this yet because I actually find it quite usefull for testing. Def something to keep in mind tho for sprint 4